refoctor: fix overlapping offset + make executor interface internal
