+++ src/main/java/com/transloadit/sdk/Transloadit.java
@@ -103,8 +103,11 @@ public class Transloadit {
     }
 
     /**
-     * Returns an AsyncAssembly instance that can be used to create a new assembly.
+     * Returns an AsyncAssembly instance that can be used to create a new assembly asynchronously.
+     * This method is good for running assemblies in the background
      *
+     * @param listener an implementation of {@link AssemblyProgressListener} to serve as a callback
+     *                 for the asynchronous assembly.
      * @return {@link AsyncAssembly}
      */
     public AsyncAssembly newAssembly(AssemblyProgressListener listener) {
@@ -114,7 +117,7 @@ public class Transloadit {
     /**
      * Returns a single assembly.
      *
-     * @param id id of the Assebly to retrieve.
+     * @param id id of the Assembly to retrieve.
      * @return {@link AssemblyResponse}
      * @throws RequestException if request to transloadit server fails.
      * @throws LocalOperationException if something goes wrong while running non-http operations.
